+++ src/frontend/apps/impress/.env
@@ -1,2 +1,3 @@
 NEXT_PUBLIC_API_ORIGIN=
 NEXT_PUBLIC_SW_DEACTIVATED=
+NEXT_PUBLIC_PUBLISH_AS_MIT=true
