+++ src/frontend/apps/impress/src/custom-next.d.ts
@@ -20,6 +20,7 @@ declare module '*.svg?url' {
 namespace NodeJS {
   interface ProcessEnv {
     NEXT_PUBLIC_API_ORIGIN?: string;
+    NEXT_PUBLIC_PUBLISH_AS_MIT?: string;
     NEXT_PUBLIC_SW_DEACTIVATED?: string;
   }
 }
